requirements.txt
certifi==2020.12.5
chardet==4.0.0
Django==2.2.2
django-ckeditor==6.0.0
django-countries==7.0
django-js-asset==1.2.2
idna==2.10
Pillow==8.1.0
pytz==2020.5
requests==2.25.1
requests-cache==0.5.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
websocket-client==0.57.0
